# codegen performance: use trim instead of lodash/trimEnd