--- a/packages/babel-generator/package.json
+++ b/packages/babel-generator/package.json
@@ -17,7 +17,8 @@
     "detect-indent": "^4.0.0",
     "jsesc": "^1.3.0",
     "lodash": "^4.2.0",
-    "source-map": "^0.5.0"
+    "source-map": "^0.5.0",
+    "trim-right": "^1.0.1"
   },
   "devDependencies": {
     "babel-helper-fixtures": "^6.22.0",