--- a/packages/babel-generator/src/buffer.js
+++ b/packages/babel-generator/src/buffer.js
@@ -1,4 +1,5 @@
 import type SourceMap from "./source-map";
+import trimRight from "trim-right";
 
 const SPACES_RE = /^[ \t]+$/;
 
@@ -41,7 +42,7 @@ export default class Buffer {
     const result = {
       // Whatever trim is used here should not execute a regex against the
       // source string since it may be arbitrarily large after all transformations
-      code: this._buf.join("").trimRight(),
+      code: trimRight(this._buf.join("")),
       map: null,
       rawMappings: map && map.getRawMappings(),
     };