@jridgewell 2017-02-02
oh, or there's https://www.npmjs.com/package/trim-right which [already](https://github.com/sindresorhus/trim-right/pull/3/files) has this optimization (from none other than @kittens).