@hzoo 2017-02-02
Wow a good find @jridgewell! https://github.com/babel/babel/issues/2008