@erikdesjardins 2017-02-02
For reference: this was originally changed from trim-right to lodash in #3500; should probably add a comment so it doesn't get switched back a second time