@danez 2017-02-02
Or maybe we can try to get this improvement into lodash? :)