@jdalton 2017-02-03
Nice find! I haven't optimized trim yet (was waiting on someone to pop-in with a scenario before I took a look at it). This looks like an easy tweak to include in future releases of Lodash.