@DrewML 2017-02-03
Thanks for doing this @jwbay. I've been super busy with work, so your contribution is very appreciated!

/cc @dmitriiabramov since Jest is waiting on this fix for a new release.