@jdalton 2017-02-03
Ah checking a bit more. This was a case where [I had previously optimized](https://github.com/lodash/lodash/blob/3.10.1/lodash.js#L643-L648) it but thought it was a bit overkill but now there's a nice scenario so :metal:.