@jdalton 2017-02-03
Out of curiosity @jwbay could you test it with [lodash.trimright](https://www.npmjs.com/package/lodash.trimright) *(the v3 version of _.trimEnd)*. It used things like `charAt` instead of index lookup for older browser support but it does a similar kind of optimization as this pr. I'm wondering if I could just revert to get the wins.