@existentialism 2017-02-03
Ran the above with `lodash.trimright` and it returns similar results:

```
5.440307 ms
0.023991 ms
0.032921 ms
0.005652 ms
0.002494 ms
0.002483 ms
0.002286 ms
0.002036 ms
0.00471 ms
0.001807 ms
0.001889 ms
0.001889 ms
0.001788 ms
0.001774 ms
0.001867 ms
0.00176 ms
0.075281 ms
0.001963 ms
0.208769 ms
0.002322 ms
```