@jdalton 2017-02-03
Thanks @existentialism!

> Ran the above with `lodash.trimright` and it returns similar results

Would you mind sanity checking the built-in `'string'.trimRight()` usage too.