@jdalton 2017-02-03
Also specify the version of Node you tested (I'm interested in v4, v6, v7)