@jdalton 2017-02-03
@jridgewell 
> Note that the new code is considerably slower when there are a large amounts of spaces at the end. We could probably speed it up further by doing a `charCodeAt` comparison so everything that matches `\s`. But this PR is enough for Babel's use case.

That's what Lodash v3 [_.trimRight](https://github.com/lodash/lodash/blob/3.0.4-npm-packages/lodash._trimmedrightindex/index.js) does.