@existentialism 2017-02-03
@jdalton 

I ran the tests on node 6.9.4.

With `String.prototype.trimRight()`:

```
4.194117 ms
0.06576 ms
0.0295 ms
0.001513 ms
0.000977 ms
0.001038 ms
0.000771 ms
0.000657 ms
0.000675 ms
0.000616 ms
0.000533 ms
0.000589 ms
0.004922 ms
0.000634 ms
0.000528 ms
0.000578 ms
0.000621 ms
0.000531 ms
0.000584 ms
0.000576 ms
```