@jdalton 2017-02-03
Great! Ya, I remember it being pretty speedy.
Here's Node v4.7.3 numbers:
```txt
27.201798 ms
0.010351 ms
0.002709 ms
0.001111 ms
0.000863 ms
0.000719 ms
0.00078 ms
0.00075 ms
0.000703 ms
0.000747 ms
0.000757 ms
0.00074 ms
0.000728 ms
0.000742 ms
0.000726 ms
0.000743 ms
0.000796 ms
0.000711 ms
0.000716 ms
0.000722 ms
```