@jdalton 2017-02-03
Looks like a quick fix (this PR) for v6 but then defer to built-in for v7 would do.