@jdalton 2017-02-03
Actually here are Node 0.12 numbers:
```txt
0.083213 ms
0.009484 ms
0.002044 ms
0.001321 ms
0.00106 ms
0.000942 ms
0.00086 ms
0.000897 ms
0.001407 ms
0.001589 ms
0.001432 ms
0.001167 ms
0.0015 ms
0.001465 ms
0.00106 ms
0.000834 ms
0.000871 ms
0.00085 ms
0.00081 ms
0.000839 ms
```
Maybe just go with the built-in all up.