@loganfsmyth 2017-02-04
Only `.trim()` is standard, `.trimRight` is a proposal that is now `.trimEnd`.

It looks like two output format tests fail if we just use `.trim()` but I'm fine if we update those tests to trim both ends.