@jwbay 2017-02-04
Ok, switched to trim. Here's the run on node 6:

```
4.412384 ms
0.034078 ms
0.03672 ms
0.002295 ms
0.002837 ms
0.002182 ms
0.001503 ms
0.001301 ms
0.001549 ms
0.001427 ms
0.001174 ms
0.001238 ms
0.001239 ms
0.001245 ms
0.001323 ms
0.001315 ms
0.001256 ms
0.001122 ms
0.001051 ms
0.001234 ms
```