@jwbay 2017-02-04
Should we go back to local impl. or use https://www.npmjs.com/package/trim-right?