@loganfsmyth 2017-02-04
Why don't we just use `trim-right` and call it good.