@jdalton 2017-02-04
Because `trimRight` is built-in and supported by all versions of Node. Why add another dep when it's completely not needed.