@loganfsmyth 2017-02-04
@jdalton It'd break the Web REPL on IE, at least according to MDN.