@loganfsmyth 2017-02-04
Babel relies on `babel-runtime` but that doesn't polyfill prototype methods, so it'd just throw.