@jdalton 2017-02-04
What is babel-runtime's browser support?
If it's not shimmable and `babel-runtime` is tied to browser support then that would be a problem.
Have you benchmarked `trim-right`?