@loganfsmyth 2017-02-04
As a library, Babel doesn't make any assumptions about non-ES5 method availability and it relies on `transform-runtime` to pick up uses of ES6 functionality and points them at `core-js`'s non-global polyfill library. `babel-runtime`'s browser support is the same as `core-js`, it just doesn't expose things that can't be detected confidently with static analysis, like prototype methods.

trim-right was linked above in https://github.com/babel/babel/pull/5255#issuecomment-276874008 as something that has this optimization.

We can also just wait if you plan to optimize this in Lodash. I have no string feelings about how we fix it.