@jdalton 2017-02-04
> As a library, Babel doesn't make any assumptions about non-ES5 method availability and it relies on `transform-runtime` to pick up uses of ES6 functionality and points them at `core-js`'s non-global polyfill library. `babel-runtime`'s browser support is the same as `core-js`, it just doesn't expose things that can't be detected confidently with static analysis, like prototype methods.

What's `core-js` browser support? I'm getting at if IE8 is involved (I hope not) but if it is then character access by index is not supported. 

> trim-right was linked above in #5255 (comment) as something that has this optimization.

Ya, it looks like it's comparable:
```txt
0.180566 ms
0.018806 ms
0.026337 ms
0.002897 ms
0.002399 ms
0.002713 ms
0.002148 ms
0.002398 ms
0.00205 ms
0.002236 ms
0.001994 ms
0.002223 ms
0.003143 ms
0.002006 ms
0.001956 ms
0.001909 ms
0.001959 ms
0.001976 ms
0.219092 ms
0.002705 ms
```

> We can also just wait if you plan to optimize this in Lodash. I have no string feelings about how we fix it.

Lodash is into it's v5 development so v4 updates are restricted to critical bugs. So this optimization won't come until v5 is released (some time away)