@loganfsmyth 2017-02-05
> What's core-js browser support? I'm getting at if IE8 is involved (I hope not) but if it is then character access by index is not supported.

Wait, IE8 doesn't support indexed string access? I'd say IE8 is beyond Babel's supported versions. I actually don't think we have an official number, but using `.trimRight` will certainly introduce something that will break on IE so avoiding it seems practical. Since `.trimRight` is not standards-track, I'm against relying on it. Does it exist in Nashorn, for example? I have no idea.

Since we can't wait on Lodash v5, and we can get acceptable performance from `trim-right`, that's my vote.